holtburger: landblock census + scenery bake 100% ACE parity hardening (full-world Phase 0-1)
Phase 0-1 of docs/full-world-bake-and-verify-plan-2026-05-30.md.

Phase 0 -- landblock-census (new bin): enumerate the content-bearing
landblocks (the bake+verify work-list) from client_cell_1.dat. A landblock
is content-bearing iff has_objects OR any non-water vertex (terrain_type
not in 0x10..=0x14, water codes per acclient.h). Of 65,025 grid LBs:
40,197 content-bearing (38,689 land + 1,508 ocean-with-structures), 26,336
ocean skipped, 3,409 dungeons (EnvCell-bearing).

Phase 1 -- scenery bake bit-parity vs ACE Scenery.Load. The prior B.5
"16,700/16,700" was only a 1e-4-tolerance match over lossy 6-decimal
strings; a new --bits diagnostic (raw f32 to_bits in scenery-bake +
scenery-cross-check, diffed by compare-bits.py) showed real <=1-ULP value
divergences on scale/qw/qz (positions always bit-exact). Fixes:
- noise.rs scale_obj: divide ratio in f32 first (matches ACE Scenery.cs:146
  maxScale/minScale), was f64 -> scale now bit-exact.
- lib.rs quaternion: compute sin/cos f64-then-narrow (closer to .NET MathF)
  -> rotation divergence 1688 -> 37/16,700.
- scenery-cross-check oracle: use the real Quaternion.CreateFromYawPitchRoll
  (MathF), not hand-rolled f64 Math.Cos/Sin, so it faithfully represents ACE.

Result: position + scale bit-exact; rotation <=1 ULP (37/16,700, ~6e-6 deg)
-- the f32-transcendental floor (.NET MathF vs Rust libm), accepted as the
parity bar. crate tests pass (incl. determinism). See the B.5 report
Phase-1 bit-identity addendum.

The --bits modes default OFF (renderer output unchanged); they + compare-bits.py
are the reusable parity harness.

+++ external/holtburger/crates/holtburger-scenery-bake/src/noise.rs
@@ -182,7 +182,14 @@ pub fn scale_obj(obj: &ObjectDesc, ix: u32, iy: u32, iq: u32) -> f32 {
     // * 2.3283064e-10
     let n = displace_noise(ix, iy, iq, 32_593);
     // pow(maxScale / minScale, n) * minScale
-    let ratio = obj.max_scale as f64 / obj.min_scale as f64;
+    //
+    // The division is done in f32 FIRST, then widened — matching ACE
+    // `Scenery.cs:146` `Math.Pow(maxScale / minScale, …)` where
+    // `maxScale`/`minScale` are `float`, so `maxScale / minScale` is an f32
+    // divide whose f32-rounded result is then promoted to double for Pow.
+    // The old code divided in f64 (`as f64 / as f64`), which keeps extra
+    // precision the f32 divide drops → ~1-ULP scale drift vs ACE.
+    let ratio = (obj.max_scale / obj.min_scale) as f64;
     (ratio.powf(n) * obj.min_scale as f64) as f32
 }
